Separate switchToPos from runtimeUpgrade.ts
will be used in other jobs in CI
